Normal udp worker takedown on timeout
We have a log line for pull data timeout. Having the other end of a pipe
stop talking to us is more expecting than needing to crash hard and
taking the whole application down. Go down normally, keep the sup
intensity to the same so we don't cover up other problems.