weekend updates (#14)
* Normal udp worker takedown on timeout

We have a log line for pull data timeout. Having the other end of a pipe
stop talking to us is more expecting than needing to crash hard and
taking the whole application down. Go down normally, keep the sup
intensity to the same so we don't cover up other problems.

* Ignore invalid DevAddr

* Drop join packets for now, we don't know what to do with them yet

* Don't crash if we can't parse netid

Kindly raise an error tuple and give everyone the option of handling it.

* bring over fix from router

This was even differenter. We don't want an binary version of the
integer, we want a 32 bit rep of the number we already know it is.

* Continue metrics on start and dump to disk on stop

* adding lager debug to test utils

* Fix failed_pull_data test

When UDP workers go down because they're not longer being responded to,
it's expected. They get shutdown as normal, which doesn't send the
'DOWN' message to monitors anymore. We still have the Pid we can check.