Apply multiplier to when we no longer consider an sc open
A state channel can get spent almost until depletion. If a restart
happens right around that time, there's a less than likely chance that
it will be spent all the way out, so will sit there as "open" refusing
to be considered "active" by the sc server. We can kick things along by
being a bit less precise about what we consider an "open" state channel.
If there's basically no DC left to spend, the sc is basically not open
and we need to open another.