handle potential for 2 listings per org
roaming-console treats the 2 net_ids that an org can have as seperate
configs that are updated with the same information. This means when
looking for join information, we may get back 2 `#eui{}` records.
They are the same, and we can ignore the second, as all we're doing is
forwarding to the roaming partner, who then get's to make the decision
which devaddr to assign.