handle hex values that are not prefixed with `0x`
behave more like Router where we always assume the dev_eui and app_eui
are hex values.

This test makes sure that valid prefixed and unprefixed hex values are
cleaned to the same number.

To remove the prefixed hex values from this repo completely would
require quite the refactor of most of the tests.