http roaming integration (#130)
* hex to binary for join accept

* pull the datarate from the payload

* match from the messages actual shape

* don't let tmst get too big

* use correct datarate type

* Store timestamp from packet instead of gateway packettime

* update regular downlink to match corrected join_accept

* was getting back negative transaction ids

* use same hex encoding for devaddr

* difference between packet time and gateway time

What we have normally called `PacketTime` is the time we received the
packet from a gateway. But there is another time inside the packet under
tmst or tmms. That field is the time the device thinks it sent the
packet. The time from the gateway is when _it_ received the packet, they
can be different. Neither can be trusted.

PacketTime (from the tmst or tmms field) is correct for queueing
downlinks.
GatewayTime is correct for giving a gregorian date for when the data was received.

* update tests for real world testing

mostly timestamp changes and hex encoding

* stop buying for no roaming agreement net ids

* fix typo

* util for truncating to 32 bit number

logging for unrecognized prstartans

* use a record for packets so we can get more dialyzer help

* update core

* make lager a top level dep

* let's try again with these traces for modules

* add protocol type to metrics

* add some logging for elli callback

otherwise elli will never log when errors are happening

* remove traces again

They didn't work because I don't have handlers for those files.
I can do those by hand right now, until they become more useful

* add util for writing current config to file