encode the dest and flow_type in the uplink token
Originally, we only had 1 configuration for a net_id so we could lookup
the sender of an xmitdata_req directly. Once that changed, I incorrectly
assumed the transactionId of an async downlink would match the uplink
transactionId. It doesn't and that makes perfect sense. The uplink token
has to be something I uplinked though, so I can encode more information
in there, namely, the place I sent the original uplink, and the
flow_type at the time of sending.